Move parseNumericString to its own file
