+++ api/src/utils/apply-query.ts
@@ -7,7 +7,6 @@ import type {
 	FieldOverview,
 	Filter,
 	NumericType,
-	NumericValue,
 	Query,
 	Relation,
 	SchemaOverview,
@@ -24,6 +23,7 @@ import { getColumn } from './get-column.js';
 import { getRelationInfo } from './get-relation-info.js';
 import { isValidUuid } from './is-valid-uuid.js';
 import { parseFilterKey } from './parse-filter-key.js';
+import { parseNumericString } from './parse-numeric-string.js';
 
 export const generateAlias = customAlphabet('abcdefghijklmnopqrstuvwxyz', 5);
 
@@ -850,27 +850,6 @@ export async function applySearch(
 			}
 		});
 	});
-}
-
-// move to its own util
-function parseNumericString(stringValue: string): NumericValue | null {
-	let number: NumericValue = Number(stringValue);
-
-	if (isNaN(number) || !Number.isFinite(number)) {
-		return null; // invalid numbers
-	}
-
-	if (number > Number.MAX_SAFE_INTEGER || number < Number.MIN_SAFE_INTEGER) {
-		number = BigInt(stringValue);
-	}
-
-	// casting parsed value back to string should be equal the original value
-	// (prevent unintended number parsing, e.g. String(7) !== "ob111")
-	if (String(number) !== stringValue) {
-		return null;
-	}
-
-	return number;
 }
 
 export function applyAggregate(
